Move createWallet to fix travis contract tests.
