+++ contracts/deploy/utils.py
@@ -1,0 +1,12 @@
+import sha3
+from ecdsa import SigningKey, SECP256k1
+from eth_utils import encode_hex
+
+
+def createWallet():
+    keccak = sha3.keccak_256()
+    priv = SigningKey.generate(curve=SECP256k1)
+    pub = priv.get_verifying_key().to_string()
+    keccak.update(pub)
+    address = keccak.hexdigest()[24:]
+    return (encode_hex(priv.to_string()), address)
